Set default model quantization to fp16 rather than fp32 to reduce memory usage
TODO:
- Add notification if content is multi-lingual to switch models (from content or obsidian settings)
- Inform user better on model choice impact (2+x slower, 2GB of RAM vs 300MB)

- Remove 1/n notes processed in bar

+++ src/infra/embedder/iframe/embeddingModel.ts
@@ -23,7 +23,7 @@ export class EmbeddingModel {
 
 		this.#pipeline = await pipeline('feature-extraction', this.config.repoId, {
 			device: this.#device,
-			dtype: webgpuAvailable ? 'fp32' : 'q8',
+			dtype: webgpuAvailable ? 'fp16' : 'q8',
 		});
 	}
